pass prep data
