+++ telco_churn_mlops/pipelines/model_builder.py
@@ -8,7 +8,6 @@ from sklearn.pipeline import Pipeline
 from sklearn.compose import ColumnTransformer
 from sklearn.impute import SimpleImputer
 from sklearn.preprocessing import FunctionTransformer, OneHotEncoder, StandardScaler
-
 from xgboost import XGBClassifier
 
 bool_features = [
