--- tests/unit/test_ab.py
@@ -1,8 +1,0 @@
-"""Unit tests for our ABTesting module"""
-
-
-def test_module():
-    """Basic sanity checks."""
-
-    from telco_churn_mlops.pipelines.ab_test import ABTestPipeline
-    pass
