+++ unit-requirements.txt
@@ -5,7 +5,7 @@ pyyaml
 pytest
 dbx
 pytest-cov
-scikit-learn
+scikit-learn==1.1.1
 mlflow
 xgboost==1.5.0
 hyperopt
